Module skyline::patching

Functions for helping patch executables

Structs

BranchBuilder
A builder type to help when replacing branches in games

Functions

patch_data⚠
Overwrite a value in read-only data with a passed value given an offset from the start of .text

patch_data_from_text⚠
Overwrite a value in read-only data with a passed value given an offset from the start of .text

patch_str⚠
Overwrite a string in read-only data with a Rust string given the offset from the start of .text

sky_memcpy⚠