Module skyline::prelude

A set of items that will likely be useful to import anyway

Designed to be used as such:

use skyline::prelude::*;

Re-exports

pub use crate::println;

Modules

vec
A contiguous growable array type with heap-allocated contents, written Vec<T>.

Macros

format
Creates a String using interpolation of runtime expressions.

vec
Creates a Vec containing the arguments.

Structs

Box
A pointer type for heap allocation.

String
A UTF-8 encoded, growable string.

Vec
A contiguous growable array type, written Vec<T> but pronounced 'vector'.

Traits

ToOwned
A generalization of Clone to borrowed data.

ToString
A trait for converting a value to a String.